requirements.txt
defopt==2.0.0
matrix-client==0.0.4
influxdb-3.0.0
